+++ docs/CHANGELOG.md
@@ -1,3 +1,4 @@
 ## Changelog for v2.1.2
 
 - Fix GitHub actions
+- Add entry to create an H2GIS in memory database